fix build failure on bad type assignment, issue #124
the internal class AsyncReceivePacket, has an ivar:
  NSMutableData *buffer;
which in practice is never mutated and the only assignments are via ummutable NSData -- clang with the error level cracked up since at least the Lion release has choked on this. the simple fix is to change the defined type to match the assignments.
http://code.google.com/p/cocoaasyncsocket/issues/detail?id=124