Add tests for reproducing the readSource freeze
Added GCDAsyncSocketReadTests with a single test that reproduces the
issue. Being async by design, the test itself is hard to write. I
created a couple of wrappers to provide a synchronous API that cleaned
up the test. We could potentially use the wrappers for other tests in
the future, if we wanted.

Since this issue requires TLS to reproduce, I've added a p12 file
containing a test certificate / private key pair. We can read that in to
simiulate an actual SSL server, and just choose not to validate it on
the client end.

To see the test fail, you can comment out the "waiting = YES;" on line
5500 of GCDAsyncSocket.m.

Tested this on the iPhone 8+ (11.2) simulator.