Refactor read tests and scaffolding
Key changes are:
- Include the the server handle along with the client and accepted handles
- Explicitly close sockets in tests instead of in deinit
- Accept before Connect
- Use random port instead of fixed `1234`

The goal of these changes was to get the newly added read test to run
in a loop of 1000 iterations. I encountered two issues with this:

1. `dispatch_source_cancel` is asynchronous which means that we can call
`socketDidDisconnect` before the file descriptor is actually released.
This causes issues because we immediately try to connect to the same
port after `socketDidDisconnect` is called. This can cause "Address
already in use" errors. We may want to see if we can wait until our
source cancellation callback has been executed before we inform the
delegate that we are disconnected. I wasn't going to make that change
here, so I worked around the issue by using a random port for each
iteration.

2. Sometimes we can still run into an issue where we timeout. Even
though the readSource is not suspended, since all of the data has
already been written, the readSource won't fire. I don't yet know where
the miscalculation is happening but we still aren't reading everything
from SSLRead. There is a workaround that I don't yet fully understand
where we can always attempt to read what's left the bytes remaining in
our read. I still want to figure out the root cause of this though.